# Integrate cudnn flash attention and add IR pass to fuse dot product attention