--- a/paddle/fluid/framework/details/build_strategy.h
+++ b/paddle/fluid/framework/details/build_strategy.h
@@ -136,7 +136,8 @@ struct BuildStrategy {
   // Fused feed forward
   bool fused_feedforward_{false};
   bool sequential_run_{false};
-
+  // Fuse dot product attention
+  bool fuse_dot_product_attention_{false};
   // mkldnn_enabled_op_types specify the operator type list to
   // use MKLDNN acceleration. It is null in default, means
   // that all the operators supported by MKLDNN will be