--- a/paddle/fluid/framework/distributed_strategy.proto
+++ b/paddle/fluid/framework/distributed_strategy.proto
@@ -171,6 +171,7 @@ message BuildStrategy {
   optional string debug_graphviz_path = 17;
   optional bool fused_attention = 18 [ default = false];
   optional bool fused_feedforward = 19 [ default = false];
+  optional bool fuse_dot_product_attention = 20 [ default = false ];
 }
 
 message ExecutionStrategy {