--- a/paddle/fluid/framework/ir/graph_pattern_detector.cc
+++ b/paddle/fluid/framework/ir/graph_pattern_detector.cc
@@ -2443,6 +2443,320 @@ PDNode *patterns::ConvElementwiseaddAct::operator()(
   return act_out;
 }
 
+PDNode *patterns::DotProductAttention::operator()(bool with_dropout) {
+  // Attention Computing
+  auto *attn_q = pattern->NewNode(attn_q_repr())
+                     ->AsInput()
+                     ->assert_is_op_output("reshape2", "Out")
+                     ->assert_is_op_input("transpose2", "X");
+  auto *attn_k = pattern->NewNode(attn_k_repr())
+                     ->AsInput()
+                     ->assert_is_op_output("reshape2", "Out")
+                     ->assert_is_op_input("transpose2", "X");
+  auto *attn_v = pattern->NewNode(attn_v_repr())
+                     ->AsInput()
+                     ->assert_is_op_output("reshape2", "Out")
+                     ->assert_is_op_input("transpose2", "X");
+
+  auto *attn_q_transpose =
+      pattern->NewNode(attn_q_transpose_repr())->assert_is_op("transpose2");
+  auto *attn_k_transpose =
+      pattern->NewNode(attn_k_transpose_repr())->assert_is_op("transpose2");
+  auto *attn_v_transpose =
+      pattern->NewNode(attn_v_transpose_repr())->assert_is_op("transpose2");
+
+  auto *attn_q_transpose_out_var =
+      pattern->NewNode(attn_q_transpose_out_repr())
+          ->assert_is_op_output("transpose2", "Out")
+          ->assert_is_op_input("scale", "X");
+  auto *attn_k_transpose_out_var =
+      pattern->NewNode(attn_k_transpose_out_repr())
+          ->assert_is_op_output("transpose2", "Out")
+          ->assert_is_op_input("matmul_v2", "Y");
+  auto *attn_v_transpose_out_var =
+      pattern->NewNode(attn_v_transpose_out_repr())
+          ->assert_is_op_output("transpose2", "Out")
+          ->assert_is_op_input("matmul_v2", "Y");
+  auto *attn_q_transpose_xshape_var =
+      pattern->NewNode(attn_q_transpose_xshape_repr())
+          ->assert_is_op_output("transpose2", "XShape");
+  auto *attn_k_transpose_xshape_var =
+      pattern->NewNode(attn_k_transpose_xshape_repr())
+          ->assert_is_op_output("transpose2", "XShape");
+  auto *attn_v_transpose_xshape_var =
+      pattern->NewNode(attn_v_transpose_xshape_repr())
+          ->assert_is_op_output("transpose2", "XShape");
+  attn_q_transpose->LinksFrom({attn_q}).LinksTo(
+      {attn_q_transpose_out_var, attn_q_transpose_xshape_var});
+  attn_k_transpose->LinksFrom({attn_k}).LinksTo(
+      {attn_k_transpose_out_var, attn_k_transpose_xshape_var});
+  attn_v_transpose->LinksFrom({attn_v}).LinksTo(
+      {attn_v_transpose_out_var, attn_v_transpose_xshape_var});
+
+  auto *attn_q_scale =
+      pattern->NewNode(attn_q_scale_repr())->assert_is_op("scale");
+  auto *attn_q_scale_out_var = pattern->NewNode(attn_q_scale_out_repr())
+                                   ->assert_is_op_output("scale", "Out")
+                                   ->assert_is_op_input("matmul_v2", "X");
+  attn_q_scale->LinksFrom({attn_q_transpose_out_var})
+      .LinksTo({attn_q_scale_out_var});
+
+  auto *attn_qk_matmul = pattern->NewNode(attn_qk_matmul_repr())
+                             ->assert_is_op("matmul_v2")
+                             ->assert_op_attr<bool>("trans_x", false)
+                             ->assert_op_attr<bool>("trans_y", true);
+  auto *attn_qk_matmul_out_var =
+      pattern->NewNode(attn_qk_matmul_out_repr())
+          ->assert_is_op_output("matmul_v2", "Out")
+          ->assert_is_op_input("elementwise_add", "X");
+  attn_qk_matmul->LinksFrom({attn_q_scale_out_var, attn_k_transpose_out_var})
+      .LinksTo({attn_qk_matmul_out_var});
+
+  auto *attn_mask_var =
+      pattern->NewNode(attn_mask_repr())->assert_is_op_input("cast", "X");
+  auto *attn_mask_cast1 =
+      pattern->NewNode(attn_mask_cast1_repr())->assert_is_op("cast");
+  auto *attn_mask_cast1_out_var = pattern->NewNode(attn_mask_cast1_out_repr())
+                                      ->assert_is_op_output("cast", "Out")
+                                      ->assert_is_op_input("cast", "X");
+  attn_mask_cast1->LinksFrom({attn_mask_var})
+      .LinksTo({attn_mask_cast1_out_var});
+
+  auto *attn_mask_cast2 =
+      pattern->NewNode(attn_mask_cast2_repr())->assert_is_op("cast");
+  auto *attn_mask_cast2_out_var = pattern->NewNode(attn_mask_cast2_out_repr())
+                                      ->assert_is_op_output("cast", "Out")
+                                      ->assert_is_op_input("scale", "X");
+  attn_mask_cast2->LinksFrom({attn_mask_cast1_out_var})
+      .LinksTo({attn_mask_cast2_out_var});
+
+  auto *attn_mask_scale1 =
+      pattern->NewNode(attn_mask_scale1_repr())->assert_is_op("scale");
+  auto *attn_mask_scale1_out_var = pattern->NewNode(attn_mask_scale1_out_repr())
+                                       ->assert_is_op_output("scale", "Out")
+                                       ->assert_is_op_input("scale", "X");
+  attn_mask_scale1->LinksFrom({attn_mask_cast2_out_var})
+      .LinksTo({attn_mask_scale1_out_var});
+
+  auto *attn_mask_scale2 =
+      pattern->NewNode(attn_mask_scale2_repr())->assert_is_op("scale");
+  auto *attn_mask_scale2_out_var =
+      pattern->NewNode(attn_mask_scale2_out_repr())
+          ->assert_is_op_output("scale", "Out")
+          ->assert_is_op_input("elementwise_add", "Y");
+  attn_mask_scale2->LinksFrom({attn_mask_scale1_out_var})
+      .LinksTo({attn_mask_scale2_out_var});
+
+  auto *attn_mask_eleadd = pattern->NewNode(attn_mask_eleadd_repr())
+                               ->assert_is_op("elementwise_add");
+  auto *attn_mask_eleadd_out_var =
+      pattern->NewNode(attn_mask_eleadd_out_repr())
+          ->assert_is_op_output("elementwise_add", "Out")
+          ->assert_is_op_input("softmax", "X");
+  attn_mask_eleadd
+      ->LinksFrom({attn_mask_scale2_out_var, attn_qk_matmul_out_var})
+      .LinksTo({attn_mask_eleadd_out_var});
+
+  auto *attn_softmax =
+      pattern->NewNode(attn_softmax_repr())->assert_is_op("softmax");
+  auto *attn_softmax_out_var = pattern->NewNode(attn_softmax_out_repr())
+                                   ->assert_is_op_output("softmax", "Out");
+  attn_softmax->LinksFrom({attn_mask_eleadd_out_var})
+      .LinksTo({attn_softmax_out_var});
+
+  auto *attn_context_matmul_input = attn_softmax_out_var;
+  if (with_dropout) {
+    attn_softmax_out_var->assert_is_op_input("dropout", "X");
+    auto *attn_dropout =
+        pattern->NewNode(attn_dropout_repr())->assert_is_op("dropout");
+    auto *attn_dropout_out_var = pattern->NewNode(attn_dropout_out_repr())
+                                     ->assert_is_op_output("dropout", "Out")
+                                     ->assert_is_op_input("matmul_v2", "X");
+    auto *attn_dropout_mask_var = pattern->NewNode(attn_dropout_mask_repr())
+                                      ->assert_is_op_output("dropout", "Mask");
+    attn_dropout->LinksFrom({attn_softmax_out_var})
+        .LinksTo({attn_dropout_out_var, attn_dropout_mask_var});
+    attn_context_matmul_input = attn_dropout_out_var;
+  } else {
+    attn_softmax_out_var->assert_is_op_input("matmul_v2", "X");
+  }
+
+  auto *attn_context_matmul =
+      pattern->NewNode(attn_context_matmul_repr())->assert_is_op("matmul_v2");
+  auto *attn_context_matmul_out_var =
+      pattern->NewNode(attn_context_matmul_out_repr())
+          ->assert_is_op_output("matmul_v2", "Out")
+          ->assert_is_op_input("transpose2", "X");
+  attn_context_matmul
+      ->LinksFrom({attn_context_matmul_input, attn_v_transpose_out_var})
+      .LinksTo({attn_context_matmul_out_var});
+
+  auto *attn_transpose =
+      pattern->NewNode(attn_transpose_repr())->assert_is_op("transpose2");
+  auto *attn_transpose_out_var = pattern->NewNode(attn_transpose_out_repr())
+                                     ->assert_is_op_output("transpose2", "Out")
+                                     ->assert_is_op_input("reshape2", "X");
+  attn_transpose->LinksFrom({attn_context_matmul_out_var})
+      .LinksTo({attn_transpose_out_var});
+  auto *attn_transpose_xshape_var =
+      pattern->NewNode(attn_transpose_xshape_repr())
+          ->assert_is_op_output("transpose2", "XShape");
+  attn_transpose->LinksFrom({attn_context_matmul_out_var})
+      .LinksTo({attn_transpose_out_var, attn_transpose_xshape_var});
+
+  return attn_transpose_out_var;
+}
+
+PDNode *patterns::DotProductAttentionGrad::operator()(bool with_dropout) {
+  auto *attn_dout_var =
+      pattern->NewNode(attn_dout_repr())
+          ->AsInput()
+          ->assert_is_op_input("transpose2_grad", GradVarName("Out"))
+          ->assert_is_op_output("reshape2_grad", GradVarName("X"));
+  auto *attn_transpose_grad = pattern->NewNode(attn_transpose_grad_repr())
+                                  ->assert_is_op("transpose2_grad");
+  auto *attn_transpose_grad_out_var =
+      pattern->NewNode(attn_transpose_grad_out_repr())
+          ->assert_is_op_output("transpose2_grad", GradVarName("X"));
+  attn_transpose_grad->LinksFrom({attn_dout_var})
+      .LinksTo({attn_transpose_grad_out_var});
+
+  attn_transpose_grad_out_var->assert_is_op_input("matmul_v2_grad",
+                                                  GradVarName("Out"));
+  auto *attn_context_matmul_grad_x_var =
+      pattern->NewNode(attn_context_matmul_grad_x_repr())
+          ->assert_is_op_input("matmul_v2_grad", "X");
+  auto *attn_context_matmul_grad_y_var =
+      pattern->NewNode(attn_context_matmul_grad_y_repr())
+          ->assert_is_op_input("matmul_v2_grad", "Y");
+  auto *attn_context_matmul_grad =
+      pattern->NewNode(attn_context_matmul_grad_repr())
+          ->assert_is_op("matmul_v2_grad");
+  auto *attn_context_matmul_grad_dx_var =
+      pattern->NewNode(attn_context_matmul_grad_dx_repr())
+          ->assert_is_op_output("matmul_v2_grad", GradVarName("X"));
+  auto *attn_context_matmul_grad_dy_var =
+      pattern->NewNode(attn_context_matmul_grad_dy_repr())
+          ->assert_is_op_output("matmul_v2_grad", GradVarName("Y"));
+  attn_context_matmul_grad
+      ->LinksFrom({attn_transpose_grad_out_var,
+                   attn_context_matmul_grad_x_var,
+                   attn_context_matmul_grad_y_var})
+      .LinksTo(
+          {attn_context_matmul_grad_dx_var, attn_context_matmul_grad_dy_var});
+
+  PDNode *attn_softmax_grad_input = nullptr;
+  PDNode *attn_softmax_out_var = nullptr;
+  if (with_dropout) {
+    auto *attn_dropout_grad = pattern->NewNode(attn_dropout_grad_repr())
+                                  ->assert_is_op("dropout_grad");
+    auto *attn_dropout_grad_out_var =
+        pattern->NewNode(attn_dropout_grad_out_repr())
+            ->assert_is_op_output("dropout_grad", GradVarName("X"));
+    attn_context_matmul_grad_dx_var->assert_is_op_input("dropout_grad",
+                                                        GradVarName("Out"));
+    attn_dropout_grad->LinksFrom({attn_context_matmul_grad_dx_var})
+        .LinksTo({attn_dropout_grad_out_var});
+    attn_softmax_grad_input = attn_dropout_grad_out_var;
+    attn_softmax_out_var = pattern->NewNode(attn_softmax_out_repr());
+
+  } else {
+    attn_context_matmul_grad_dx_var->assert_is_op_input("softmax_grad",
+                                                        GradVarName("Out"));
+    attn_softmax_grad_input = attn_context_matmul_grad_dx_var;
+    attn_softmax_out_var = attn_context_matmul_grad_x_var;
+  }
+  attn_softmax_out_var->assert_is_op_input("softmax_grad", "Out");
+
+  auto *attn_softmax_grad =
+      pattern->NewNode(attn_softmax_grad_repr())->assert_is_op("softmax_grad");
+  auto *attn_softmax_grad_out_var =
+      pattern->NewNode(attn_softmax_grad_out_repr())
+          ->assert_is_op_output("softmax_grad", GradVarName("X"));
+  attn_softmax_grad->LinksFrom({attn_softmax_out_var, attn_softmax_grad_input})
+      .LinksTo({attn_softmax_grad_out_var});
+
+  attn_softmax_grad_out_var->assert_is_op_input("elementwise_add_grad",
+                                                GradVarName("Out"));
+  auto *attn_mask_eleadd_grad_mask_var =
+      pattern->NewNode(attn_mask_eleadd_grad_mask_repr())
+          ->assert_is_op_input("elementwise_add_grad", "Y");
+  auto *attn_mask_eleadd_grad = pattern->NewNode(attn_mask_eleadd_grad_repr())
+                                    ->assert_is_op("elementwise_add_grad");
+  auto *attn_mask_eleadd_grad_dx_var =
+      pattern->NewNode(attn_mask_eleadd_grad_dx_repr())
+          ->assert_is_op_output("elementwise_add_grad", GradVarName("X"));
+  attn_mask_eleadd_grad
+      ->LinksFrom({attn_softmax_grad_out_var, attn_mask_eleadd_grad_mask_var})
+      .LinksTo({attn_mask_eleadd_grad_dx_var});
+
+  attn_mask_eleadd_grad_dx_var->assert_is_op_input("matmul_v2_grad",
+                                                   GradVarName("Out"));
+  auto *attn_qk_matmul_grad_x_var =
+      pattern->NewNode(attn_qk_matmul_grad_x_repr())
+          ->assert_is_op_input("matmul_v2_grad", "X");
+  auto *attn_qk_matmul_grad_y_var =
+      pattern->NewNode(attn_qk_matmul_grad_y_repr())
+          ->assert_is_op_input("matmul_v2_grad", "Y");
+  auto *attn_qk_matmul_grad = pattern->NewNode(attn_qk_matmul_grad_repr())
+                                  ->assert_is_op("matmul_v2_grad");
+  auto *attn_qk_matmul_grad_dx_var =
+      pattern->NewNode(attn_qk_matmul_grad_dx_repr())
+          ->assert_is_op_output("matmul_v2_grad", GradVarName("X"));
+  auto *attn_qk_matmul_grad_dy_var =
+      pattern->NewNode(attn_qk_matmul_grad_dy_repr())
+          ->assert_is_op_output("matmul_v2_grad", GradVarName("Y"));
+  attn_qk_matmul_grad
+      ->LinksFrom({attn_mask_eleadd_grad_dx_var,
+                   attn_qk_matmul_grad_x_var,
+                   attn_qk_matmul_grad_y_var})
+      .LinksTo({attn_qk_matmul_grad_dx_var, attn_qk_matmul_grad_dy_var});
+
+  attn_qk_matmul_grad_dx_var->assert_is_op_input("scale", "X");
+  auto *attn_scale_grad =
+      pattern->NewNode(attn_scale_grad_repr())->assert_is_op("scale");
+  auto *attn_scale_grad_out_var = pattern->NewNode(attn_scale_grad_out_repr())
+                                      ->assert_is_op_output("scale", "Out");
+  attn_scale_grad->LinksFrom({attn_qk_matmul_grad_dx_var})
+      .LinksTo({attn_scale_grad_out_var});
+
+  attn_scale_grad_out_var->assert_is_op_input("transpose2_grad",
+                                              GradVarName("Out"));
+
+  // q -> transpose2_grad -> reshape2_grad
+  auto *attn_q_transpose_grad = pattern->NewNode(attn_q_transpose_grad_repr())
+                                    ->assert_is_op("transpose2_grad");
+  auto *attn_dq = pattern->NewNode(attn_dq_repr())
+                      ->assert_is_op_output("transpose2_grad", GradVarName("X"))
+                      ->assert_is_op_input("reshape2_grad", GradVarName("Out"));
+  attn_q_transpose_grad->LinksFrom({attn_scale_grad_out_var})
+      .LinksTo({attn_dq});
+
+  // k -> transpose2_grad -> reshape2_grad
+  attn_qk_matmul_grad_dy_var->assert_is_op_input("transpose2_grad",
+                                                 GradVarName("Out"));
+  auto *attn_k_transpose_grad = pattern->NewNode(attn_k_transpose_grad_repr())
+                                    ->assert_is_op("transpose2_grad");
+  auto *attn_dk = pattern->NewNode(attn_dk_repr())
+                      ->assert_is_op_output("transpose2_grad", GradVarName("X"))
+                      ->assert_is_op_input("reshape2_grad", GradVarName("Out"));
+  attn_k_transpose_grad->LinksFrom({attn_qk_matmul_grad_dy_var})
+      .LinksTo({attn_dk});
+
+  // v -> transpose2_grad -> slice_grad
+  attn_context_matmul_grad_dy_var->assert_is_op_input("transpose2_grad",
+                                                      GradVarName("Out"));
+  auto *attn_v_transpose_grad = pattern->NewNode(attn_v_transpose_grad_repr())
+                                    ->assert_is_op("transpose2_grad");
+  auto *attn_dv = pattern->NewNode(attn_dv_repr())
+                      ->assert_is_op_output("transpose2_grad", GradVarName("X"))
+                      ->assert_is_op_input("reshape2_grad", GradVarName("Out"));
+  attn_v_transpose_grad->LinksFrom({attn_context_matmul_grad_dy_var})
+      .LinksTo({attn_dv});
+
+  return attn_dq;
+}
+
 PDNode *patterns::VitAttention::operator()(PDNode *in) {
   in->AsInput();
   std::unordered_set<std::string> matmul_ops{"matrix_multiply"};