… paddle/fluid/framework/details/CMakeLists.txt
… paddle/fluid/framework/ir/CMakeLists.txt
… paddle/fluid/framework/ir/fuse_dot_product_attention_pass.cc
… paddle/fluid/pir/dialect/op_generator/ops_api_gen.py
… paddle/fluid/pir/transforms/fusion/fused_dot_product_attention_pass.cc
… paddle/fluid/pybind/parallel_executor.cc
… paddle/phi/api/yaml/fused_backward.yaml
… paddle/phi/api/yaml/fused_ops.yaml
… paddle/phi/infermeta/backward.h
… paddle/phi/infermeta/fusion.cc
… paddle/phi/infermeta/fusion.h
… paddle/phi/kernels/CMakeLists.txt
… paddle/phi/kernels/fusion/gpu/fused_dot_product_attention_op.cu
… paddle/phi/kernels/gpudnn/mha_cudnn_frontend.cu
… paddle/phi/kernels/gpudnn/mha_cudnn_frontend.h
… python/paddle/distributed/passes/cpp_pass.py
… python/paddle/framework/ir.py
… python/paddle/incubate/nn/functional/__init__.py
… python/paddle/incubate/nn/functional/fused_dot_product_attention.py
… test/ir/pir/fused_pass/test_fused_dot_product_attention_pass.py
… test/legacy_test/CMakeLists.txt
… test/legacy_test/test_fuse_dot_product_attention_pass.py
… test/legacy_test/test_fused_dot_product_attention_op.py
… tools/gpups_test.sh