### PR types
New features

### PR changes
OPs

### Description
CUDNN v8 provide fused (flash) attention kernel to accelerate TransformerLayer(refer to [cudnn-frontend/fused_mha_sample](https://github.com/NVIDIA/cudnn-frontend/blob/main/samples/f16_flash_mha_sample.cpp)). To be exact, it only fuse the calculation of scaled dot product attention, which does not include qkv projection and output projection, but only includes scale, bmm1, mask, softmax, dropout, and bmm2. So, let's call it fused_dot_product_attention instead of fused_multi_head_attention.

1. Added the following OPs to compute **_scaled dot product attention_**  (forward + backward):
    - fused_dot_product_attention/fused_dot_product_attention_grad
2. Added a new IR pass to detect above scaled dot product attention pattern.
3. Added a class member fuse_dot_product_attention to BuildStrategy to enable fuse_dot_product_attention_pass.

### TransformerLayer perf test
GPU: A100
config: batch size=32, hidden size=1024, seq len=1024
| original (time per step) |  fuse dot product attention (time per step) | speedup|
| ------- | ------- | -------|
|   576 ms |   406 ms |  1.40x |


**Build strategy usage sample (CUDNN >= 8.9.6 is required):**
```
class TransformerLayer(paddle.nn.Layer):
    def __init__(self, hidden, num_heads):
        super().__init__()
        self.encoder_layer = paddle.nn.TransformerEncoderLayer(hidden, num_heads, 4*hidden, dropout=0.1)
        self.encoder = paddle.nn.TransformerEncoder(self.encoder_layer, 2)

    def forward(self, q, mask):
        out = self.encoder(q, mask)
        loss = paddle.mean(out)
        return loss

paddle.enable_static()
main_prog = paddle.static.Program()
startup_prog = paddle.static.Program()

batch_size = 32
head_size = 64
num_head = 12
hidden = head_size * num_heads
seq_len = 128

with paddle.static.program_guard(main_prog, startup_prog):
    enc_input = paddle.static.data(name='enc_input', shape=[-1, -1, hidden], dtype="float32")
    attn_mask = paddle.static.data(name='mask', shape=[-1, 1, seq_len, seq_len], dtype="int32")
    model = TransformerLayer(hidden, num_heads)
    loss = model(enc_input, attn_mask)
    opt = paddle.optimizer.SGD(learning_rate=0.1)
    amp_list = paddle.static.amp.CustomOpLists(custom_white_list=['softmax'])
    opt = paddle.static.amp.decorate(
        optimizer=opt,
        amp_lists=amp_list,
        init_loss_scaling=128.0,
        use_dynamic_loss_scaling=True)
    opt.minimize(loss)


place = paddle.CUDAPlace(0)
exe = paddle.static.Executor(place)
exe.run(startup_prog)

build_strategy = paddle.static.BuildStrategy()
build_strategy.fuse_dot_product_attention = True
main_prog = paddle.static.CompiledProgram(main_prog)
main_prog = main_prog.with_data_parallel(loss_name=loss.name,
                                         build_strategy=build_strategy,
                                         places=place)
exe.run(main_prog, feed=feed)
# 2 subgraphs be fused to fused_dot_product_attention op
# 2 subgraphs be fused to fused_dot_product_attention_grad op
```