@Wong4j 2023-11-08
# IR pass
IR pass部分的细节较多，这里做一些解释以便review

## 设计思路:

**QKV输入**：fused_dot_product_attention 不计算qkv projection和out projection，只计算中间的scale, bmm1, mask, softmax, dropout, bmm2等。输入是q, k, v的projection的结果（已经reshape成[b, s, h, d]）和mask。

我们用IR pass来对subgraph进行遍历和替换，将scaled dot product attention的所有op和中间变量都匹配出来，将其替换成fused op，如下图，左边是原始的subgraph，紫色的部分会被替换成右图中的绿色的fused op。
![image](https://github.com/PaddlePaddle/Paddle/assets/21985950/e0cc9148-f38c-483e-b630-a811f6afd0c2)


Note: 
- 由于cuDNN flash attention并不能直接输入mask，而是输入actual_seqlen，所以我额外加了一个cuda kernel来将mask转换成actual_seqlen。
- cuDNN flash attention 支持causal mask，并且我在fused_dot_product_attention op中加入了is_causal_masking的参数。但nn.MultiHeadAttention中没有相关参数，理论上IR pass中可以根据mask的排布来确定是否是causal mask，但这并不是一个好的方法。所以目前的设计只能用来处理普通的padding mask。

## 实现细节：

**fuse_dot_product_attention_pass.h:** 
**fuse_dot_product_attention_pass.cc**

1. **3个OpCache**
    1. MaskCache: 用来存mask节点），比如BERT large有24个layer，对同一个输入batch来说，计算24次mha的fwd/bwd都会使用同一组mask，所以这个mask需要存起来，作为替换subgraph之后的fused op的fwd/bwd的输入。
    2. OutputCache：cudnn mha中bwd的计算需要fwd的变量，包含SoftmaxStats，fwd output和rng state。所以fwd的时候要存储在cache中，然后bwd再取出来用。
    3. QKVCache：同理，fwd计算时将qkv存储在cache中，在对应的bwd计算中取出来，用来计算dq, dk, dv。
2. **nodes_to_remove**
    如上所述，以BERT-large为例，需要将24个fwd/bwd中所有的subgraph都替换成fused op。由于某些node的指针存储在cache中，在bwd时要再次使用，所以不能每次替换一个subgraph后就立即删除掉被替换的节点，而是替换完所有的subgraph后再删除。