@zhiqiu 2023-11-14
We are now developing a new ir, called PIR. Would you mind transfer the fusion pass to PIR? You can refer to [fused_gemm_epilogue_pass](https://github.com/PaddlePaddle/Paddle/pull/58188).