@zyfncg 2023-11-27
> > We are now developing a new ir, called PIR. Would you mind transfer the fusion pass to PIR? You can refer to [fused_gemm_epilogue_pass](https://github.com/PaddlePaddle/Paddle/pull/58188).
> 
> @zhiqiu 请教一下，我参考其他的PR，在PIR下面添加了fusion pass，但单测却没有成功将subgraph替换成fused op，也没有任何错误信息。请问有没有什么debug技巧呢？老的IR pass可以打开GLOG_v来看subgraph匹配每一步的细节，但是新的IR我不知道如何debug。

新fusion pass的匹配逻辑在`paddle/fluid/pir/drr/drr_rewrite_pattern.cc`中，使用GLOG_v可以打印出一些信息，如果不够可以手动添加GLOG信息用于定位