@Wong4j 2023-12-04
> We are now developing a new ir, called PIR. Would you mind transfer the fusion pass to PIR? You can refer to [fused_gemm_epilogue_pass](https://github.com/PaddlePaddle/Paddle/pull/58188).

@zyfncg 我已经在PIR下添加了这个pass和单测，但是似乎新的IR pass需要用PassManager run来apply，这样的话能放到build strategy里面来开启吗？
因为我设计的其中一个功能就是想要仅添加一行代码`build_strategy.fuse_dot_product_attention=True`，来自动fuse模型中的attention。