@Wong4j 2023-12-05
@zyfncg 这个PR目前包含新旧两套IR的实现，因为我们(JoC BERT)[https://github.com/NVIDIA/DeepLearningExamples/tree/master/PaddlePaddle/LanguageModeling/BERT]的代码目前还是用旧的静态图写法，依赖旧的IR。所以我觉得这个PR可以先审核完合入，我后续研究一下engine，更新我们的BERT代码，然后再提新的PR完全迁移到PIR。