@Wong4j 2023-12-07
- @onecatcn PR-CI-APPROVAL和PR-CI-APPROVAL需要以下approve:
0. one RD (XiaoguangHu01, jeff41404, lanxianghit or qingqing01) approval for API change.
and one TPM approval for API change: jzhang533/ZhangJun, sunzhongkai588/SunZhongKai, Ligoml/LiMengLiu for general APIs.
1. one TPM approval for API documents change: jzhang533/ZhangJun, sunzhongkai588/SunZhongKai, Ligoml/LiMengLiu for general API docs.
2. print or std::cout is not recommended for direct use, please use logging or VLOG. If it is necessary to use, please contact tianshuo78520a (Recommend) or zhangbo9674 review and approve.
3. must have raindrops2sea or XiaoguangHu01 approval for change 20+ files or add than 1000+ lines of content.
4. one RD (XiaoguangHu01,chenwhql,zhiqiu,Xreki,luotao1,qili93,Aurelius84) approval for the usage of const_cast.
5. Unittest is not allowed to be disabled. You must have one RD (kolinwei(Recommend), wanghuancoder, luotao1, QingshuChen, qili93 or ZzSean or Aurelius84) approval for the usage of @unittest.skip or @unittest.skipIf.

- PR-CI-Coverage：我新加的单测是在A100+cudnn 8.9.6的机器上跑的，在PR-CI-GpuPS中都pass了，都能覆盖到Coverage中没有跑到的code，所以这种情况应该可以豁免？@tianshuo78520a